feat[#31] : useSnippetStore 생성
- useInteractionStore 제거하고 useSnippetStore 내부로 모두 통합
