--- app/src/features/snippet/model/useInteractionStatusStore.ts
@@ -1,10 +1,0 @@
-import { createStore } from "@/shared/lib/store";
-
-type Status = "loading" | "idle" | "succeed" | "fail";
-interface InteractionStatusStore {
-  status: Status;
-}
-
-export const useInteractionStatusStore = createStore<InteractionStatusStore>({
-  status: "idle",
-});
